Created surface_t using common members from both planes and wallseg sections.
Missing texture replacement is now done once when needed after a call to R_UpdateSector instead of each frame. As this change is now higher up the logic tree surfaces with missing textures are now processed for fakeradio, shiny textures etc using the replacements.

Fixed a bug with CcmdMovePlane which resulted in a seg violation.