Shuffled data structures in renderer to simplify usage and to allow for future enhancement (such as an unlimited number of planes per sector and eventually, slopes). RP_QUAD rendpolys and wall vissprites are now handled using four vertices, which now have a Z coordinate.
RP_FLAT rendpolys (plane) vertices now also use a Z coordinate. All Planes of a subsector now share the same vertexes[XY] (which have been moved to subsectorinfo_t) and the plane's height is set in RL_PreparePlane.

Fixed various small glitches as and when found.

TODO: Reorganise RP_QUAD vertices so that WallSegs can be rendered as triangle strips (and for neatness).