New busy mode. The main engine thread loops in a drawer, which updates the screen. In the background, the worker thread does the actual task of loading a map, loading textures, starting up the engine (not quite yet, though), connects to a server, etc.
Texture uploading was revised to make it possible for the worker thread to upload textures in a deferred mode, where the main thread actually does the real upload. The main thread thus has control of exactly how and when the textures get uploaded. There is a timeout to prevent frames from being too long.

There may be some issues left with texture parameter setting, especially related to mipmapping.

The next step: modify engine startup so that GL is inited as early as possible to allow for graphics to be drawn while the startup worker boots the engine. After that, an actual implementation for Con_BusyDrawer().