Fixed: After a mimic sector, it was not possible for that sector to then support a special thinker of any description (if the source of the mimic had a thinker at the time P_CopySector() is called). Basically, do NOT copy xsector->specialdata.
This poses the question: when mimicking a sector, should any active special thinkers be copied also? Currently the engine-side speed and target properties for the planes ARE copied but since there is no thinker for the mimic target, they will never move.

If so, we need to handle this correctly (stopping any existing thinker and then create a copy of the one used for the source of the mimic). This in turn poses another question: Should the properties of the new thinker be re-evaluated at the time of the mimic or should they be direct copies of those of the mimicked thinker (this could lead to floors/ceilings moving onto the other side of one another)?

Any thoughts?