Use floating point 0 - 1 when dealing with rgb colors. All code updated except when dealing directly with vertices (i.e. in the bias lighting code).
Found and fixed an issue with the light grid which resulted in evaluated points returning {0, 0, 0} if the evaluation happens after a change but before an update. In this instance we copy the old value prior to being marked for update (not entirely accurate but given that the next update is only a single tic away; its OK).

Part #2