Worked on the texture management a bit.
We now have a hash for patches uploaded to the videocard.
Revised flat management so that they are indexed by id and not lump numbers. During init (and update) we loop through all lumps between F_Start/F_End and add the flats to a LUT and assign ids.

The members of dpatch_t are now in host byte order (converted in R_CachePatch()).

Next step: concatenate flats, textures and sprite ids into one continues index to simplify texture management and unify for net games.

TODO: fix issues with savegames and converting from the v1 save format which used lump indices.

Part 4 of 4