Continued work on the texture management code.
Revised how textures loaded from raw lumps are managed. These plus the recent changes to the way patches and flats are managed meant that it is no longer necessary to maintain lumptexinfo entries for ALL loaded lumps, so they have been removed.

TODO: Strangely, there now seems to be a visible seam on the jHexen titlescreen. This is odd because the manipulation code hasn't been changed. Neverless it will need to be fixed.