Adjusted the map data object structures to remove some redundancies.
When the automap "show subsectors" cheat is enabled; draw "minisegs" too (note, these are only present in GL node data).

TODO: We really need to get things so that ALL subsectors have a complete set of segs (forming a convex polygon). It would make life so much easier and the data could be organised into a more "render-ready" structure.