Replicate the edge fade effect seen in the original games' automap using multitexturing.
TODO: Still needs a bit of tweaking. Either I'm doing something wrong or DGL_MODULATE_TEXTURE=12 doesn't work as I expect. Also, there seems to be a small amount of panning which is scrolling the mask in screen-space, got to remove that and then clamp the edges of the mask texture.

All <gamename>.pk3s need to be rebuilt due to the addition of mapmask.lmp