Fixed issues with runtime textures (flats and patches) not being cleared after a texreset.
Implemented runtime interfaces to control the bpp mode and whether the window runs full screen (cvars, ccmds and control panel).

Under Win32, intercept the WM_SIZE:SIZE_MAXIMIZED message sent by the system (the shell is about to maximize the window) and instead, instigate the change ourselves via a call to GL_ChangeResolution() and automatically switch to fullscreen mode.

Cleaned up drOpenGL ready for future development.

TODO: There are some remaining issues with texreset. Most important is the issue with texture translations; because we now store a ptr to the current translation in surface_t->material (to avoid searching each frame) these become invalid after a (tex)reset. At the end of a (texreset) we need to update all map surfaces but this in turn relies on knowing when a map is actually loaded. We should implement a method to unload the current map and record the state, engine-side.