Completed removal of the old start up window.
Implemented a wrapper for window management. Currently, only the required system window code has been implemented for win32, on Unix and Mac OSX we should either use a library or implement code which uses X11 directly.

Fixed issues with crosshairs not centered to the view window correctly.

Renamed DD_SCREEN* constants to DD_WINDOW* to avoid confusion with the current video mode.

When changing videomode's in DG_CreateContext() do NOT automatically change the window size/style under win32.

Removed the titlebar drawn in the control panel.

Fixed mouse cursor scaling in the UI to be independent of window aspect ratio and to scale relatively to window dimensions.

Fixed scaling issues in the UI due to inconsistencies in coordinate space usage (some places used 800x600 others 1000x1000). Replaced functions for converting relative to screen space coordinates with versions using equations which better avoid 32bit integer overflow (issues experienced in super hi resolutions).

Parametrized numerous drawing routines to allow them to be reused and to minimize calls to the window management wrappers to determine window properties.

Slight changes to busy mode so that it can be used when we don't have a window to draw in.

Todo: Continue factoring host system window management out of DG_CreateContext() and into our wrapped window code. There is an issue with switching between windowed and fullscreen under win32 with the window position being set back to [0,0]. There seems to be a scaling issue in the control panel/UI; not all objects are scaling relatively (list box scrollbar, button borders and more).