Attempt at fixing a segfault in the particle generator management after an engine "reset". The generator definitions are re-read during the reset and so ptcgen_t->def ptrs are invalid. Currently, only damage and type generators are updated. Map generators are simply destroyed as we have no means to identify them once the definitions have been re-read. Flat triggered generators are simply destroyed as they will be respawned anyway within a few tics by P_CheckPtcPlanes().
TODO: This doesn't seem to be working. The segfault is prevented but it appears that all generators are being destroyed instead of updated where possible. I suspect the problem to be in determining the generator type in p_particle.c around line #1422. Skyjake - can you take a look at this?