Cleanup and updated style of drOpenGL. Re-jigged things slightly under win32.
TODO: Under WIN32, OpenGL extension function pointers are only valid for the pixel format in effect for the rendering context (inherited from the device context) at the time of query. If we create additional windows and/or change the pixel format at any time the function ptrs must be re-acquired. This is the direct opposite of what happens under X11 where extension ptrs are pixel format independent. So under WIN32, we need to keep a set of ptrs for each rendering context and make sure we update when the rendering context changes, through wglMakeCurrent().