SDL-based window management;
Begun re-factoring things so that the same code can be used for all platforms. This presents some issues under win32 which will need to be resolved (for example it is currently not possible to build with the SDL-based version due to additional changes required in the renderer (init/rendering context stuff)).

Does this build and run for you skyjake?