Default bindings for basic player movement. Turning, walking, sidestep, and lookdir working correctly.
Attempting to get the player movement working as it used to. Ticcmd_t will be removed, but for now we can put the new control state values into it and see that everything works as expected.

Experiencing odd stutter in the Z-direction (turn off view bobbing to see it better), need to investigate why... Maybe related to the fractional G_Ticker.