Jaakko's crusade for Usability of APIs continues: I made the (IMO) exceedingly cumbersome Sys_GetWindowDimensions() function unnecessary with a global pointer to the current ddwindow_t. To access the details of the window where drawing is currently being done, just use e.g., "theWindow->width", or "theWindow->bpp".
The requirement to call a 5-parameter function, which may on top of that fail for reasons unknown, necessitating error handling in functions that really shouldn't even be called if there is no window, was the main hurdle for me.

If on Win32 it is necessary to know the dimensions of the window before the window has actually been created (a situation which would be best remedied by creating the window earlier), one can always init theWindow and mainWindow with the expected values even though an actual physical window is not yet available.  

Also some reorganization of Doomsday headers: mainly that sys_* headers shouldn't be included from de_base.h, but from de_system.h.