Further optimized renderWalls() to reduce the number of DMU calls. This includes using P_SubsectorBoxIterator() to reduce the number of subsectors in the set that must be drawn to those (possibly) within the map's view frame (instead of attempting to draw the lot).
Still, the greatest cause of slowdown when drawing the automap is due to the number of DMU calls and the associated performance penalties. We'll either need to speed up DMU or look to implement a more clever optimization for automap drawing.

Fixed autmap mask texture coordinate scaling when the map window is not located at [0, 0].

Fixed the automap zoom speed cvar. These should be read from the player's cfg and applied to the cfg held in the player's automap.

Added cvar "map-pan-speed": automap panning speed.
Added cvar "map-pan-resetonopen": reset the automap pan position when opening the map (default is ON; original DOOM behaviour).