Fixed HQ2x scaling to work on big-endian processors.
Magic tricks to improve readability of the original font lumps. When the font lumps are being uploaded, the DD_UPSCALE_AND_SHARPEN_PATCHES variable is set to "true", which causes the patches to be HQ2x-upscaled and a black outline will be applied to the upscaled version of the texture. Gives a marked improvement on Heretic and Hexen fonts, but Doom fonts could use a contrast enhancement, too.