Split the code for the concepts of A) luminous object in the world (lumobj_t) and B) Projected light on a surface (dynlight_t) into separate source files in preparation for further work.
Enhanced the dynlight linked with an option to link into a surface light list so as they are sorted by brightness at insertion time. This removes the need to iterate the dynlight list of surface to find the brightest light when converting a rendpoly to a vissprite (for masked/alpha'd middle textures). Also fixed a couple of lighting bugs.