Implemented an interface for setting the global game action so that we have a chance to react to changes.
Grab the new exports from the dgl plugin (we really need to remove/rethink this abstraction).

Numerous changes to the games to allow starting levels with no things (players are spawned as cameras at [0,0]).

Continued cleaning up the game code. Begun separating the code of the menu functionality and the game-specific parts. Fixed various bugs in the menu.

Looking back at my early code, it is almost painful. No doubt I'll spend a day or so soon and completely reimplement the menu as it should have been all along.

Part 4.