Added new cached map file reader/writer.
Once a map has been loaded by a map converter and reconstructed engine-side via the runtime map editing interface, it is written to the cached map folder for future use.

Thats the plan at least. It isn't yet fully implemented. Also, after discussion with Yagisan, we need to replace the use of lzss with zlib.