Fixed: Fairly frequently, the vertex coordinates generated for fakeradio edge shadowpolys would result in "folds" at the inner offset. This was due to us storing two separate offsets for each side of the edge and due to numerical inaccuracies; these would sometimes end up not being equal. This has been fixed by using only one set of offsets for each pair of edges and storing them in a the vertex lineowner ring for easy access.
Todo: This has the shortcoming that all edge pairs need to have a set of shadow offsets allocated for them as opposed to only those that actually receive edge shadows. However, given that we are now using roughly half the memory of the previous schema,this is still a big improvement.