Added symbolic events and parameter substitution in event binding commands.
Control doubleclicks are tracked and produce a corresponding symbolic event. A command can be bound to this symbolic event. The symbolic event contains the originating player control number, which can be used in the bound command with the %i or %p sequences.

This is used currently only for doubleclick-positive-walk, but symbolic events could very well have other uses.