Continued working on the controls and bindings.
A special mode was added to the engine called "symbolic echo", which will cause input events to be converted to textual symbolic events. This is for the binding grabber, and only enabled when an event is being grabbed.

Still no binding deletion, although all the infrastructure is now in place as the Delete function was added as a flag to the menus.