Extensive changes to the renderer to support multiple local players and to reduce the amount of work done per viewport (shared between multiple viewers where possible). Part 3 of 3.
Apologies for the size and complexity of the changes in this one "mega" commit. I lost my internet connection for a nearly a week, which whilst it allowed me to spend much more time on deng then I would normally; resulted in a complex mass of changes that I couldn't see an easy way to pick appart into logically separateable pieces. Sorry.