Fixed All Games: Due to the different approach taken by Doomsday and DOOM.exe in defining texture/flat animations and the numerous PWADS which exist that exploit DOOM.exe behaviour; it is not possible to use Group definitions to externalize the IWAD animations. All default Group definitions have been removed and instead the default animations are registered by each game during startup (and engine reset) in P_InitPicAnims. This fixes the long-standing problem of PWADS which abused the DOOM.exe behaviour not working correctly when played with jDoom/jHeretic.
Note: This fix depends upon an assumption that will need to be resolved; Materials for textures defined in TEXTURE(1) lumps are created in the same order in which they are defined and thus their indices are the same (relatively) to one another. This is due to there being no current means to resolve a material num given a texture id.