Fixed use of uninitialized color in LG_Evaluate when a given block has no sector.
Todo: This doesn't fix the core of the problem though. The issue is that the sector coverage determination done during initialization of the lightgrid is incorrectly determining that some blocks are not covered by any sector. This is most likely due to a numerical round-off problem but the question remains, is this current algorithm *really* what we want?