Improved event handling in non-game states. Paged menus now scroll item-by-item, rather than by page.
The "gameui" binding context was added. Contexts may now specify a fallback responder, which gets to process the event if the context doesn't handle it. The "gameui" context makes sure the menu gets opened in non-game states.

PageUp and PageDown can now be used for scrolling the menus by page. Page scrolling was changed so that the cursor is always centered on the screen and the items scroll one by one. It could be further improved with smooth scrolling, but that's actually a lot of work to implement, given how the drawing is currently done.