Fixed (all games): Bug #2431808 Unknown pickup items leading to a "P_SpecialThing: Unknown gettable thing" critical error. Rather than terminate, handle this gracefully by treating the item as inert (removing the mobj from world, counting bonus and executing any scripted special but doing nothing further). Fixed (jDoom/jHeretic/jDoom64): Bug 2530948 Giving armour via XG's ltc_power line type would only work as expected if the player had already picked up an armor item.
In the process of fixing the above I have re-implemented the mechanism for picking up items in jHexen. All games should be updated to use this new code (then commonized) as it fixes a number of issues in original Heretic and Hexen and should prove to be more easily maintainable.