Various changes to the format of ACS data written to/read from saved games. No fix for the jHexen floor waggles yet though.
Incremented the jHexen save version. Split IO for the global ACS data store and world vars into new routines, added a segment marker and version byte for this data. Incremented the version byte for the global ACS data store segment in order to avoid writing 240 bytes of unnecessary junk. Fixed a couple of 64bit portability issues in the process.