Fixed: (DOOM.EXE bug) if the game is saved while a repeatable switch is momentarily in the "on" position, upon loading the saved game the switch would never return to the "off" position. Fixed: Material changes on sidedef surfaces as a result of XG's automatic switch swapping not working and switches defined via the SWITCHES lump ignored.
In fixing the above, the old DOOM button_t stuff has been removed. We now make use of the thinker system with a new thinker T_MaterialChanger, which will change the material of a given surface in X tics time. This thinker is written to and restored from saved games.