Fixed: "Dynamic lights bleeding through solid walls" see here (http://sourceforge.net/tracker/?func=detail&aid=2717071&group_id=74815&atid=542099).
Redesigned algorithm for spreading objlinks to generate objlink > subsector contacts. The old algorithm worked on sectors, spreading over linedef boundaries and then linked the objlink for all subsectors in the contacted sectors. The new algorithm is subsector based and spreads accross segs with a partner/back seg.

Todo: At some point we should document all of this objlink spreading, surface contact and surface list stuff. I'm quite sure that there is some redundancy going on here and with the aid of some top-down design documentation, I'm sure we could come up with a better design to eliminate it.