Fixed: "jHexen: Savegame Crash with jXRP" see here (http://sourceforge.net/tracker/?func=detail&aid=2647053&group_id=74815&atid=542099). Added: cvar "rend-dev-generator-show-indices" - Display particle generator origin indices (for debug). Note that type-triggered and plane-triggered generators are not handled as these are inherently origin-less.
In fixing the above I have implemented a new concept: Public/Private Thinkers.
When a thinker is added to the thinker lists, the engine decides whether this thinker is visible publicly, via the API's thinker interface routines. This visibility then determines which thinker list the thinker is added to within the engine. Only publicly-visible thinkers can be interacted with via the public API.

Particle generator thinkers (P_PtcGenThinker) are engine-internal and are not made public. This means that when the game asks to re-init the thinker lists (typically after loading a saved game) these thinkers will not be affected.

Todo: It would be possible to utilize the same system for other kinds of thinkers that the game does not need to know about. I'm thinking client-side mobjs and plane/poly movers.