Peppered the main loop with profile timers enabled when DD_PROFILE is defined.
I had a feeling something was going awry and performance was gradually degrading. I made a 15min run of the starting area of NDCP2.zip::MAP01 with no addons, god mode enabled and the player backed into a corner with the view direction, angle etc constant. The results in doomsday.out certainly suggest there is a problem somewhere as by the end of the run, per-frame operations were taking noticeably longer to complete.