Fixed: "Doom Armour Icon on full screen is squashed vertically" (see here http://sourceforge.net/tracker/?func=detail&aid=2762310&group_id=74815&atid=542099).
Reverted changes to ST_drawHUDSprite which assumed that the engine did not have to resize the HUD sprite textures to pow2 dimensions. As use of the NPOT extension is disabled by default this will suffice for now, however, we should address this properly once the HUD sprites are wraped by gltexture_t.