"import record" statement.
"import record" allows copying a module's namespace into the local
namespace as a local record. This will make it possible to serialize
the imported namespace as part of the record. A plain "import" will
only create a variable that references the imported namespace.

Should be helpful with Config and Map/World/Object records.
I still need to implement serialization of Functions, though.