Binding modifiers.
Replaced the Strafe control with generic modifier controls.
dd_share.h currently defines 4 modifiers. The old strafe control
is now CTL_MODIFIER_1.

Sidestepping and turning are now defined with bindings instead of
with player movement logic.

The controls menu still hasn't been updated with this in mind.
Possibly the binding grabber should detect the modifiers separately
so that it would be possible to bind, e.g., to "Alt + Left" (when
Alt is the bound modifier).