Fixed a crash with Hexen's "Random" player class.
Evidently the use of the enum type was causing trouble with gcc.
When queried about pClass, gdb indicated that it was "optimized away
by the compiler", even though that clearly shouldn't be done. The
crash occurred because the value of pClass ended up being negative
regardless of the if statement.