Added: Smoothing of per-linedef lightlevel angle deltas. Approximated rounded surfaces are lit more realistically when using the DOOM lighting model. Added: CVar "rend-light-wall-angle-smooth" 1= Enable wall angle lightlevel delta smoothing (enabled by default). Changed: Default value for "rend-light-wall-angle" is now 1.2f as the new smoothing makes this less noticeable.
Whilst this smoothing algorithm catches 90% of cases where it should occur there is one case that it does not. The case it doesn't handle is the same as the one case that fakeradio ommits (where a single-sided linedef is colinear with a two-sided linedef whose back sector is open). This last case cannot really be solved with the current stratergy and requires a redesign of the seg drawing loop in Rend_RenderSubsector (which I have already begun in the mapcache branch). As such I decided to leave it as is for now.

I have been sitting on this idea for a while. Now that it is clear that a Beta6.9 release is necessary - I decided to quickly implement it for this release.