When in a transition is in progress do not call the play tickers (in DD_Ticker).
Whilst in busy mode the deui bindcontext is now active. This prevents game-context input events from accumulating while busy. If a transition is in effect the deui bindcontext remains active until it ends else it deactivates when the busy worker thread returns. A bit of a kludge? Instead of using the deui context should we instead have a special-case "ignore all" context (not a real context, nothing can be bound to it)?