Reordered G_DoLoadMap so that if a briefing is not present; the current music track is changed and instantly paused, just before entering busy mode to load the map. The music is then unpaused when the map starts.
The idea is to reposition the very noticeable system locking which occurs under Windows when a music track is changed (culprit unknown at present). Ideally the cached music file write should be done during busy mode but thats not an option presently due to multi-thread communication with the audio system (and thus the underlying API).

The result is much smoother transitions when loading maps on Windows.

Note however that this means there will be no music playing during the load process (not such a big deal? Shouldn't take long anyway, especially once map caching is implemented).