Reordered G_DoCompleted; the current music track is changed and instantly paused, just before entering busy mode to prepare the intermission. The music is then unpaused when the intermission starts.
The idea is to reposition the very noticeable system locking which occurs under Windows when a music track is changed (culprit unknown at present). Ideally the cached music file write should be done during busy mode but thats not an option presently due to multi-thread communication with the audio system (and thus the underlying API).

The result is much smoother transitions when loading maps on Windows.