Fixed All games: "General: Mobjs overlapping more than one sector". (see here http://sourceforge.net/tracker/?func=detail&aid=2855894&group_id=74815&atid=542099)
Revised the fix applied in beta6.6 for the "Bullets not visible in map cheat" issue. Although we now want all mobjs added to the blockmap it is not correct that all mobjs with MF_SOLID are considered so everywhere. This means that we still need the MF_NOBLOCKMAP flag as it defines slightly different behaviour compared to !MF_SOLID. We still ignore the MF_NOBLOCKMAP flag as far as linking goes however it is now checked for in the iteration methods which were previously sensitive to this (namely, PIT_ChangeSector).

Ensure you exectute packres.py as the Thing and Flag definitions have changed.