Fixed bug 2849414: Shoots after loading/saving from menu.
There already was a mechanism that kept track which binding context
each individual input device (a particular key/axis/hat) was
associated with. However, when this association changes, the old
mechanism did not react in any way. Now I've added a logic wherein
the state of a particular device key/axis/hat is flagged as "expired"
when the context association changes while the device is not in its
default state.

In practice, we consider the state of a device unknown when it changes
its binding context until we can be sure that it has really returned
to its default state.