Fixed bug 2899607: Double press input events.
The event sequence responder was eating events that were part of a
partial sequence. This prevented the execution of the commands
bound the to eaten events (e.g., pause for P).

I don't think it's possible for the event sequence responder to
eat events during the forming of a sequence without affecting
regular bindings.