Began clean up of GL_DrawText in preparation for implementing a proper visible dimension calculator for Text. * Split the paramater block parsing out of GL_DrawText. * Added drawtextstate_t and separated initialization and call-time "defaults".
Fixed Common code library: When drawing paramater text blocks as used with the Patch Replacement system, the "break" directive was interpreted immediately when encountered. However if the same paramater block also made changes to font or scale before the use of the break directive; the subsequent break(s) would use the new font and/or scale rather than that which was in effect before the paramater block began.