Reimplemented the Text shadow effect used in the menu in order to eliminate references to the menu in gl_drawcompositefont.c
gl_drawcompositefont.c is now ready for intergration into the engine. Initially, I'll move a copy of gl_drawpatch.c into the engine for use with this. The next step will surely be redesigning how the character patch graphics are loaded and drawn (use an atlus and generate a set of display lists for each character).