Fixed handful of uninitialized assignments shown up by static analysis. Thanks for the report yagisan.
All but two of these were innoxious however as they are not read after the fact. The rest of the issues are false positives due to the logic flow algorithm used by the analysis tool being unaware that execution stops after a call to Con_Error, which never returns.

We should consider enhancing Con_Error so as to allow the analysis tool(s) to detect the termination of the logic flow.