Addressed "Sectors with zero linedefs result in SIGSEGV after load" (see here: http://sourceforge.net/tracker/?func=detail&atid=542099&aid=3190384&group_id=74815). * Fixed: Recalculation of the sector environmental sound properties (e.g., after a plane move) did not work correctly (resulting in accumulation instead). * Fixed: Map boundary calculation failed to consider sectors with no linedefs. Possibly resulting in the inclussion of point [X=0, Y=0] regardless whether any "real" geometry exists there. * Fixed: Flat-triggered particle generators were being unnecessarily spawned for sectors with no linedefs. * Various other related minor fixes.
I've left things as-is on game-side for now, as it is very likely that these "null sectors" have an effect on mod compatibility. Logically they should be omitted from structures like the taglists but DOOM does not. Similarly, XG fails to consider them also and will reference them when conditions are right. What to do?