Added DOOMSDAY_RELEASE_TYPE and a more logical version ID.
The Doomsday version text now contains the release type
(unstable, candidate, stable). Revised the way the version
text is generated so it's more logical. Replaced the old
"DGL" (Doomsday GL) with "OpenGL" as the renderer implementation
in the version ID.

The platform_release script reads the release type, but it
doesn't currently affect the build in any way. Ultimately it
will have an effect on the produced file name and where the
file is copied in the apt repository.

Also, added headers to engine sources in CMakeLists.txt so
that they appear in Qt Creator as part of the project.