Client: Fixing mobj/clmobj flags
Trying to make sense of when flags are set and unset. Added
assertions for the solid and remote flags. The client is now
able to move after respawning.

The logic is still that the real mobj is solid and not remote, and
handled by the client's game logic, while the clmobj is not solid
and remote, and updated by the server when necessary.