Mobj flags synced from server to client
In an effort to ensure that mobj state is properly transmitted from
the server to the clients, moved the mobj flags and health from
game-side to engine-side. This means the engine has visibility
to them, but doesn't know what the values mean -- they are simply
copied to the clients.

Also, the mobj Z data in the delta now includes floor and ceiling Z
heights. This fixes a problem with mobjs not knowing their proper
plane heights until they moved.