Client sends accurate position w/momentum
Previously the client was sending position coordinates
only as (short) integers, without the fractional part. This
caused the server to think that the client was sometimes
slightly inside a wall. Now the accurate client position
is sent to the server, complete with momentum so that
other clients will be able to accurately model the
movement path of remote players.